+++ noodles-tabix/src/lib.rs
@@ -45,6 +45,10 @@ pub use self::io::Writer;
 pub use self::r#async::Reader as AsyncReader;
 
 #[cfg(feature = "async")]
+#[deprecated(
+    since = "0.45.0",
+    note = "Use `noodles_tabix::r#async::io::Writer` instead."
+)]
 pub use self::r#async::Writer as AsyncWriter;
 
 use std::{fs::File, path::Path};